+++ packages/primitives/src/bnHexWei/utils/isToBN.ts
@@ -1,0 +1,5 @@
+import { IToBN } from '../types';
+
+export function isToBN(value: unknown): value is IToBN {
+  return typeof value === 'object' && !!value && 'toBN' in value;
+}